refactor: journals dto와 type 분리
