+++ backend/src/journals/types/journals-list-types.ts
